# api/api.yaml
openapi: 3.0.0
info:
  title: Todo API
  description: Todo API
  version: "0.1"
servers:
  - url: http://localhost:8080
paths:
  /api/v1/todo/{id}:
    get:
      summary: Get todo
      operationId: getTodo
      tags:
        - todo
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Todo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/OperationFailed"
    delete:
      summary: Delete todo
      operationId: deleteTodo
      tags:
        - todo
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Todo was deleted
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/OperationFailed"
  /api/v1/todo:
    get:
      summary: List todos
      operationId: listTodos
      tags:
        - todo
      responses:
        "200":
          description: Todo
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Todo"
        default:
          $ref: "#/components/responses/OperationFailed"
    post:
      summary: Create todo
      operationId: createTodo
      tags:
        - todo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTodoRequest"
      responses:
        "201":
          description: Todo was created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateTodoResponse"
        default:
          $ref: "#/components/responses/OperationFailed"
    delete:
      summary: Delete all todos
      operationId: deleteAllTodos
      tags:
        - todo
      responses:
        "204":
          description: All todos were deleted
        default:
          $ref: "#/components/responses/OperationFailed"
components:
  responses:
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    OperationFailed:
      description: Operation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    Todo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        content:
          type: string
      required:
        - id
        - title
        - content
    CreateTodoRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 20
        content:
          type: string
      required:
        - title
        - content
    CreateTodoResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
      required:
        - id
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        debug:
          type: string
      required:
        - message
        - debug