# services/todo/server/server.go
package server

import (
	"database/sql"
	"errors"
	"fmt"
	"net/http"

	"github.com/google/uuid"

	"github.com/shaxbee/todo-app-skaffold/pkg/api"
	"github.com/shaxbee/todo-app-skaffold/pkg/httperror"
	"github.com/shaxbee/todo-app-skaffold/pkg/httprouter"
	"github.com/shaxbee/todo-app-skaffold/services/todo/model"
)

type TodoServer struct {
	queries *model.Queries
}

func New(db model.DBTX) *TodoServer {
	return &TodoServer{
		queries: model.New(db),
	}
}

func (s *TodoServer) RegisterRoutes(router *httprouter.Router) {
	router.Handler(http.MethodPost, "/api/v1/todo", s.Create)
	router.Handler(http.MethodGet, "/api/v1/todo/:id", s.Get)
	router.Handler(http.MethodGet, "/api/v1/todo", s.List)
	router.Handler(http.MethodDelete, "/api/v1/todo/:id", s.Delete)
	router.Handler(http.MethodDelete, "/api/v1/todo", s.DeleteAll)
}

func (s *TodoServer) Create(w http.ResponseWriter, req *http.Request) error {
	ctx := req.Context()

	var ctReq api.CreateTodoRequest
	if err := httprouter.JSONRequest(req, &ctReq); err != nil {
		return err
	}

	if len(ctReq.Title) > 20 {
		return httperror.New(http.StatusBadRequest, httperror.Message("title should have maximum length of 20 characters"))
	}

	id, err := uuid.NewRandom()
	if err != nil {
		return fmt.Errorf("failed to generate todo id: %w", err)
	}

	err = s.queries.Create(ctx, model.CreateParams{
		ID:      id,
		Title:   ctReq.Title,
		Content: ctReq.Content,
	})
	if err != nil {
		return fmt.Errorf("failed to create todo: %w", err)
	}

	return httprouter.JSONResponse(w, http.StatusCreated, api.CreateTodoResponse{
		Id: id,
	})
}

func (s *TodoServer) Get(w http.ResponseWriter, req *http.Request) error {
	ctx := req.Context()

	rawID := httprouter.ParamsFromContext(ctx).ByName("id")

	id, err := uuid.Parse(rawID)
	if err != nil {
		return httperror.New(http.StatusBadRequest, httperror.Message("invalid id"), httperror.Cause(err))
	}

	t, err := s.queries.Get(ctx, id)

	switch {
	case errors.Is(err, sql.ErrNoRows):
		return httperror.New(http.StatusNotFound, httperror.Messagef("todo %q not found", id))
	case err != nil:
		return fmt.Errorf("failed to get todo: %w", err)
	}

	return httprouter.JSONResponse(w, http.StatusOK, api.Todo{
		Id:      t.ID,
		Title:   t.Title,
		Content: t.Content,
	})
}

func (s *TodoServer) List(w http.ResponseWriter, req *http.Request) error {
	ctx := req.Context()

	todos, err := s.queries.List(ctx)
	if err != nil {
		return fmt.Errorf("failed to list todos: %w", err)
	}

	resTodos := make([]api.Todo, len(todos))
	for i, t := range todos {
		resTodos[i] = api.Todo{
			Id:      t.ID,
			Title:   t.Title,
			Content: t.Content,
		}
	}

	return httprouter.JSONResponse(w, http.StatusOK, resTodos)
}

func (s *TodoServer) Delete(w http.ResponseWriter, req *http.Request) error {
	ctx := req.Context()

	rawID := httprouter.ParamsFromContext(ctx).ByName("id")

	id, err := uuid.Parse(rawID)
	if err != nil {
		return httperror.New(http.StatusBadRequest, httperror.Message("invalid id"), httperror.Cause(err))
	}

	n, err := s.queries.Delete(ctx, id)
	switch {
	case err != nil:
		return fmt.Errorf("failed to delete todo: %w", err)
	case n == 0:
		return httperror.New(http.StatusNotFound, httperror.Messagef("todo %q not found", id))
	default:
		w.WriteHeader(http.StatusNoContent)
		return nil
	}
}

func (s *TodoServer) DeleteAll(w http.ResponseWriter, req *http.Request) error {
	if err := s.queries.DeleteAll(req.Context()); err != nil {
		return fmt.Errorf("failed to delete all todos: %w", err)
	}

	w.WriteHeader(http.StatusNoContent)

	return nil
}